Made the code way more robust and perform 2 points better on avg
The evaluation of the Individual is done only when the fitness is used somewhere, python style
Removed the isSorted bool, it only makes the code harder to maintain and ads only a couple of ms of run time, but the budget is 12s
Switched the new offspring generation back to the simple crossover: resulted in a 2 point increase on avg, over 10 random samples (new method avg=5.9; crossover 7.96)
Fixed several bugs, such as parentsRatio not being properly assigned, taking twice the request number of parents etc.
Doubled the mutation rate from 1% to 2%; increased swing from 0.01 maximum to 0.1.
Commented some lines from player32, so that the errors don't pester us. They will be uncommented in later versions